add: create Explanations index page
this started out life as a single internal reference document of a few terms, starting about six weeks back. the more it developed, the more it seemed to make for a singular glossary document.

however, the first 60 seconds of that endeavor made clear this was—necessarily—a comprehensive, fractal glossary of core concepts