add: `P` directory, `Phase Transition` and `Protocol`
really excited about phase transition in particular. put a lot of time into the descriptions for reader resonance, and i'm realizing that the "Mathematical Context" section is an opportunity to *format* things like formal math, but use the space to *explain* the tensor calculus in ways that can resonate with average minds. at one point it dawned on me: 'what's the point of just restating the formalism i spent all that time on in /math/? 90% of people don't give a shit about the math at all, and the 10% that will, only will if it starts out recognizable and relatable';  also tried to write all the other explanations as coherently as i could for average people, or better yet if i was trying to explain it to my mom